back/requirements.txt
Django==4.1
uwsgi==2.0.20
PyMySQL==1.0.2
djangorestframework==3.13.1
django-filter==22.1
django-allauth==0.51.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.12
Pillow==9.3.0
celery==5.2.7
django-celery-results==2.4.0
django-redis==5.2.0